requirements.txt
# Dependencies for Spotify Ads Mute
pycaw>=20230407
pywin32>=306
comtypes>=1.2.0
psutil>=5.9.0

# Dependencies for System Tray version
pystray>=0.19.0
Pillow>=10.0.0
